Cumulative commit: - Changed vmMain interface: the command parameter is now int instead of intptr_t to match the datatypes of the arguments sent. - Changed GametypeCommand interface: use intptr_t for the parameters and return type (instead of int) to match the datatypes of the arguments sent. The command parameter remains int.

+++ code/game/g_main.c
@@ -235,7 +235,7 @@ This is the only way control passes into the module.
 This must be the very first function compiled into the .q3vm file
 ================
 */
-Q_EXPORT intptr_t vmMain( intptr_t command, intptr_t arg0, intptr_t arg1, intptr_t arg2, intptr_t arg3, intptr_t arg4, intptr_t arg5,
+Q_EXPORT intptr_t vmMain( int command, intptr_t arg0, intptr_t arg1, intptr_t arg2, intptr_t arg3, intptr_t arg4, intptr_t arg5,
                           intptr_t arg6, intptr_t arg7, intptr_t arg8, intptr_t arg9, intptr_t arg10, intptr_t arg11 )
 {
     switch ( command )
